requirements.txt
Django>=4.2
djangorestframework>=3.14
django-environ>=0.12
celery>=5.3
redis>=4.5
requests>=2.31